Combined CertificateBatchLogRepository with BatchLogRepository
