+++ src/SFA.DAS.AssessorService.Application/Handlers/Certificates/GetCertificatesForBatchNumberHandler.cs
@@ -8,16 +8,16 @@ namespace SFA.DAS.AssessorService.Application.Handlers.Certificates
 {
     public class GetCertificatesForBatchNumberHandler : IRequestHandler<GetCertificatesForBatchNumberRequest, CertificatesForBatchNumberResponse>
     {
-        private readonly ICertificateBatchLogRepository _certificateBatchLogRepository;
+        private readonly IBatchLogQueryRepository _batchLogQueryRepository;
         
-        public GetCertificatesForBatchNumberHandler(ICertificateBatchLogRepository certificateBatchLogRepository)
+        public GetCertificatesForBatchNumberHandler(IBatchLogQueryRepository batchLogQueryRepository)
         {
-            _certificateBatchLogRepository = certificateBatchLogRepository;
+            _batchLogQueryRepository = batchLogQueryRepository;
         }
 
         public async Task<CertificatesForBatchNumberResponse> Handle(GetCertificatesForBatchNumberRequest request, CancellationToken cancellationToken)
         {
-            var certificates = await _certificateBatchLogRepository.GetCertificatesForBatch(request.BatchNumber);
+            var certificates = await _batchLogQueryRepository.GetCertificatesForBatch(request.BatchNumber);
             
             var certificatesToBePrintedResponse = new CertificatesForBatchNumberResponse()
             {
