+++ src/SFA.DAS.AssessorService.Application/Interfaces/IBatchLogQueryRepository.cs
@@ -1,4 +1,6 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using SFA.DAS.AssessorService.Domain.DTOs;
 using SFA.DAS.AssessorService.Domain.Entities;
 
 namespace SFA.DAS.AssessorService.Application.Interfaces
@@ -7,8 +9,12 @@ namespace SFA.DAS.AssessorService.Application.Interfaces
     {
         Task<BatchLog> Get(int batchNumber);
 
+        Task<CertificateBatchLog> GetCertificateBatchLog(int batchNumber, string certificateReference);
+
         Task<BatchLog> GetLastBatchLog();
 
         Task<int?> GetBatchNumberReadyToPrint();
+
+        Task<List<CertificatePrintSummary>> GetCertificatesForBatch(int batchNumber);
     }
 }
